tests [make]: created testing for if blocks
If blocks not test. This includes `if`, `if..else`, and `else if`.
Tests include testing of conditional evaluation. Block parsing,
whether or not propogation within the block work, and general
behavior. Additionally minor changes were made to various existing
tests. Though as v1 rolls out, many more tests need to be created
especially as coverage is lacking barely at 60%